Move types from VendorDashboard to types file
