+++ src/components/App/App.tsx
@@ -19,9 +19,7 @@ import VendorOrders from '../VendorOrders/VendorOrders';
 import CustomerOrders from '../CustomerOrders/CustomerOrders';
 import CustomerSettings from '../CustomerSettings/CustomerSettings';
 import VendorSettings from '../VendorSettings/VendorSettings';
-import { NewItem, Item } from '../VendorDashboard/VendorDashboard';
-import { Customer } from '../../types';
-import { Vendor } from '../VendorLogIn/VendorLogIn';
+import { Customer, Vendor, NewItem, Item } from '../../types';
 import Map from '../Map/Map';
 import ErrorPage from '../ErrorPage/ErrorPage';
 import 'leaflet/dist/leaflet.css';
