+++ src/components/VendorLogIn/VendorLogIn.tsx
@@ -1,18 +1,18 @@
-import React, { useState } from "react";
-import "./VendorLogIn.css";
-import { useNavigate } from "react-router";
-import { VendorDashboardProps } from "../VendorDashboard/VendorDashboard";
+import React, { useState } from 'react';
+import './VendorLogIn.css';
+import { useNavigate } from 'react-router';
+import { VendorDashboardProps, Vendor } from '../../types';
 
-export type Vendor = {
-  email: string;
-  password: string;
-  id: number;
-  first_name: string;
-  last_name: string;
-  vendor_name: string;
-  market: number;
-  location: string;
-};
+// export type Vendor = {
+//   email: string;
+//   password: string;
+//   id: number;
+//   first_name: string;
+//   last_name: string;
+//   vendor_name: string;
+//   market: number;
+//   location: string;
+// };
 
 const VendorLogIn = ({
   allItems,
@@ -23,11 +23,11 @@ const VendorLogIn = ({
 }: VendorDashboardProps) => {
   const navigate = useNavigate();
   function handleGoBack() {
-    navigate("/");
+    navigate('/');
   }
 
   function handleSignup() {
-    navigate("/vendorsignup");
+    navigate('/vendorsignup');
   }
 
   function handleLogin(e: any) {
@@ -42,20 +42,20 @@ const VendorLogIn = ({
       setIsVendor(true);
       setCurrentUserId(validUser.id);
       navigate(`/vendordashboard/${validUser.id}`);
-      setMessage("");
+      setMessage('');
     } else {
       setMessage(
-        "Oops! Check your email and password. Or sign up for an account."
+        'Oops! Check your email and password. Or sign up for an account.'
       );
       setTimeout(() => {
-        setMessage("");
+        setMessage('');
       }, 3500);
     }
   }
 
-  const [vendorLoginEmail, setVendorLoginEmail] = useState("");
-  const [vendorLoginPassword, setVendorLoginPassword] = useState("");
-  const [message, setMessage] = useState("");
+  const [vendorLoginEmail, setVendorLoginEmail] = useState('');
+  const [vendorLoginPassword, setVendorLoginPassword] = useState('');
+  const [message, setMessage] = useState('');
   return (
     <form className="vendor-login-container">
       <h2 className="vendor-login-header">MARKET NEXT DOOR</h2>
@@ -67,7 +67,7 @@ const VendorLogIn = ({
         name="vendorEmail"
         placeholder="Enter email..."
         value={vendorLoginEmail}
-        onChange={(e) => setVendorLoginEmail(e.target.value)}
+        onChange={e => setVendorLoginEmail(e.target.value)}
       />
       <input
         className="vendor-login-input"
@@ -75,15 +75,12 @@ const VendorLogIn = ({
         name="vendorPassword"
         placeholder="Enter password..."
         value={vendorLoginPassword}
-        onChange={(e) => setVendorLoginPassword(e.target.value)}
+        onChange={e => setVendorLoginPassword(e.target.value)}
       />
-      <p className="message" style={{ fontSize: "1.4rem" }}>
-        {message}{" "}
+      <p className="message" style={{ fontSize: '1.4rem' }}>
+        {message}{' '}
       </p>
-      <button
-        className="vendor-login-submit-btn"
-        onClick={(e) => handleLogin(e)}
-      >
+      <button className="vendor-login-submit-btn" onClick={e => handleLogin(e)}>
         SIGN IN
       </button>
       <p className="vendor-login-text">OR</p>
